app/requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
python-dotenv
aiohttp
websockets
openai
python-multipart